+++ util-crypto/scrypt/fromU8a.ts
@@ -1,7 +1,7 @@
 
 import type { ScryptParams } from './types.ts';
 
-import { u8aToBn } from 'https://deno.land/x/polkadot@0.2.44/util/mod.ts';
+import { u8aToBn } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { BN_LE_OPTS } from '../bn.ts';
 import { DEFAULT_PARAMS } from './defaults.ts';