+++ util-crypto/sha/asU8a.ts
@@ -1,8 +1,8 @@
 
-import { sha256 as sha256Js } from 'https://esm.sh/@noble/hashes@1.3.2/sha256.js';
-import { sha512 as sha512Js } from 'https://esm.sh/@noble/hashes@1.3.2/sha512.js';
+import { sha256 as sha256Js } from 'https://esm.sh/@noble/hashes@1.3.3/sha256.js';
+import { sha512 as sha512Js } from 'https://esm.sh/@noble/hashes@1.3.3/sha512.js';
 
-import { sha256, sha512 } from 'https://deno.land/x/polkadot@0.2.44/wasm-crypto/mod.ts';
+import { sha256, sha512 } from 'https://deno.land/x/polkadot/wasm-crypto/mod.ts';
 
 import { createBitHasher, createDualHasher } from '../helpers.ts';