+++ util-crypto/sr25519/vrfVerify.ts
@@ -1,6 +1,6 @@
 
-import { u8aToU8a } from 'https://deno.land/x/polkadot@0.2.44/util/mod.ts';
-import { vrfVerify } from 'https://deno.land/x/polkadot@0.2.44/wasm-crypto/mod.ts';
+import { u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
+import { vrfVerify } from 'https://deno.land/x/polkadot/wasm-crypto/mod.ts';
 
 const EMPTY_U8A = new Uint8Array();