+++ util-crypto/xxhash/asU8a.ts
@@ -1,6 +1,6 @@
 
-import { hasBigInt, u8aToU8a } from 'https://deno.land/x/polkadot@0.2.44/util/mod.ts';
-import { isReady, twox } from 'https://deno.land/x/polkadot@0.2.44/wasm-crypto/mod.ts';
+import { hasBigInt, u8aToU8a } from 'https://deno.land/x/polkadot/util/mod.ts';
+import { isReady, twox } from 'https://deno.land/x/polkadot/wasm-crypto/mod.ts';
 
 import { createAsHex } from '../helpers.ts';
 import { xxhash64 } from './xxhash64.ts';
@@ -14,7 +14,7 @@ import { xxhash64 } from './xxhash64.ts';
  * <BR>
  *
  * ```javascript
- * import { xxhashAsU8a } from 'https://deno.land/x/polkadot@0.2.44/util-crypto/mod.ts';
+ * import { xxhashAsU8a } from 'https://deno.land/x/polkadot/util-crypto/mod.ts';
  *
  * xxhashAsU8a('abc'); // => 0x44bc2cf5ad770999
  * ```