+++ util/hex/hasPrefix.ts
@@ -12,7 +12,7 @@ import { isHex } from '../is/hex.ts';
  * <BR>
  *
  * ```javascript
- * import { hexHasPrefix } from 'https://deno.land/x/polkadot@0.2.44/util/mod.ts';
+ * import { hexHasPrefix } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * console.log('has prefix', hexHasPrefix('0x1234')); // => true
  * ```