+++ util/is/number.ts
@@ -8,7 +8,7 @@
  * <BR>
  *
  * ```javascript
- * import { isNumber } from 'https://deno.land/x/polkadot@0.2.44/util/mod.ts';
+ * import { isNumber } from 'https://deno.land/x/polkadot/util/mod.ts';
  *
  * console.log('isNumber', isNumber(1234)); // => true
  * ```