+++ x-randomvalues/react-native.ts
@@ -2,8 +2,8 @@
 
 import { NativeModules } from 'https://esm.sh/react-native';
 
-import { base64Decode } from 'https://deno.land/x/polkadot@0.2.44/wasm-util/base64.ts';
-import { xglobal } from 'https://deno.land/x/polkadot@0.2.44/x-global/mod.ts';
+import { base64Decode } from 'https://deno.land/x/polkadot/wasm-util/base64.ts';
+import { xglobal } from 'https://deno.land/x/polkadot/x-global/mod.ts';
 
 import { crypto as cryptoBrowser, getRandomValues as getRandomValuesBrowser } from './browser.ts';
 import { insecureRandomValues } from './fallback.ts';